requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.29.2
diffusers==0.15.0
xformers==0.0.20
accelerate==0.19.0
safetensors==0.4.2
huggingface-hub==0.20.3
einops==0.8.1
Pillow==9.4.0
opencv-python==4.9.0.80
tqdm==4.65.0
numpy==1.24.4
scipy==1.11.4
PyYAML==6.0         
omegaconf==2.3.0    
ptflops==0.7.4      
thop==0.1.1.post2209072238  
